requirements.txt
streamlit
openai
python-dotenv
pandas
numpy
PyPDF2
reportlab